refactor: WC-219 extract breadcrumbs
